+++ Modules/Intent.Modules.Common.AI/release-notes.md
@@ -1,3 +1,7 @@
 ### Version 1.0.0
 
-- `IntentSemanticKernelFactory` for creating SemanticKernel `Kernel` instances based on user settings.
+- New Feature: `IntentSemanticKernelFactory` for creating SemanticKernel `Kernel` instances based on user settings. 
+
+> [!NOTE]
+>
+> Supports the following AI Providers: OpenAI, Azure OpenAI and Anthropic.